nodeTag is size_t in MSH4.1

+++ meshio/msh_io/msh4_1.py
@@ -359,7 +359,7 @@ def _write_elements(fh, cells, write_binary):
             )
             fh.write(numpy.array([node_idcs.shape[0]], dtype=c_ulong).tostring())
 
-            assert node_idcs.dtype == c_int
+            assert node_idcs.dtype == c_size_t
             data = numpy.column_stack(
                 [
                     numpy.arange(
